requirements.txt
Flask
Flask-Cors
requests
beautifulsoup4
selenium
webdriver-manager
python-whois
joblib
numpy
scikit-learn
